+++ der2rsa.c
@@ -116,7 +116,7 @@ int
 rsa_keypair_from_der(struct rsa_public_key *pub,
 		     struct rsa_private_key *priv,
 		     unsigned limit, 
-		     unsigned length, const uint8_t *data)
+		     size_t length, const uint8_t *data)
 {
   struct asn1_der_iterator i;
   enum asn1_iterator_result res;
